Changed the XML machine description for the AccessSpec Tag Patterns.  LLRP defines two tag patterns name TagPattern1 and TagPattern2.  These are the only named parameters within the LLRP specification and create some trouble for some of the library implementations.
After discussion with the group, it was agreed that we'd represent the two tag patterns as a list of 1-N tag patterns.  This makes a cleaner API as the two tag patterns are identical and the spec doesn't dictate any order requirements.

The Name atttribute for a parameter is still a valid part of the schema.  This change is made to ease the naming burden that this put on some libraries.