constructors of LLRPMessage and decode/encodeBinary and decode/encodeXML binary methods now throw the InvalidLLRPMessage Exception
in the decode/encodeBinary methods, I added catch statements to capture IllegalArgumentExceptions and MissingParameterExceptions generated by parameters and field of the message. (since decode/encodeXML are implemented in the subclasses, the corresponding catch statements for the XML impl can be found in the message templates (LLRP/custom).

I removed the toString method.

The toBinaryString and to toXMLString methods catch InvalidLLRPMessageExceptions and return an error message as the string.