added a new copy statement that copies an extension XSD to the org/llrp/ltk package so that it is included in the jar. The extension XSD needs to be specified as a property at the beginning of the build file.
Note: This is a suboptimal solution because only a single extension can be specified and the same properties need to be specified in the build.xml AND in generator.properties. We are planning on changing this but this is not straightforward since ant only supports java.util.Properties and we are using org.apache.commons.configuration for generator.properties.