we made the following changes to the LTKJava networking package - the API remains almost identical.
LLRPAcceptor and LLRPConnector now check the ConnectionAttemptStatus of the READER_EVENT_NOTIFICATION message sent by the reader and throw a LLRPConnectionAttemptFailedException in case of unsuccessful status (only change to API).

LLRPIoHandlerAdapter is no longer an interface but implements IoHandlerAdaptor. It also replies to incoming KEEP_ALIVE messages and captures the ConnectionAttemptStatus in READER_EVENT_NOTIFICATION messages.

LLRPIoHandlerAdapterImpl was removed. The functionality is now available in LLRPIoHandlerAdaptor.